refactor push route
